feat: implement CSV data selector
