+++ src/components/CSVExport/CSVExport.js
@@ -7,25 +7,7 @@ import SaveAltIcon from '@material-ui/icons/SaveAlt';
 // Modules
 import { CSVLink } from 'react-csv';
 
-// Constants
-import { MAP } from 'config/default';
-
-export const CSVExport = ({ data, hiddenColumns }) => {
-  const csvData = data.map(item => {
-    const filteredItem = {};
-    Object.keys(item).forEach((key, index) => {
-      if (!MAP[index].isHidden && !hiddenColumns[key]) {
-        if (typeof item[key] === 'object') {
-          filteredItem[key] = Object.values(item[key]).join(' ');
-        } else {
-          filteredItem[key] = item[key];
-        }
-      }
-    });
-
-    return filteredItem;
-  });
-
+export const CSVExport = ({ csvData }) => {
   return (
     <Grid item>
       <CSVLink
@@ -44,11 +26,9 @@ export const CSVExport = ({ data, hiddenColumns }) => {
 };
 
 CSVExport.propTypes = {
-  data: PropTypes.arrayOf(PropTypes.object),
-  hiddenColumns: PropTypes.arrayOf(PropTypes.object),
+  csvData: PropTypes.arrayOf(PropTypes.object),
 };
 
 CSVExport.defaultProps = {
-  data: [],
-  hiddenColumns: [],
+  csvData: [],
 };
